Sorting (U): add nan check
This is a crucial feature that's absent from most Pareto
Sort implementations, and it causes all manner of grief.
Because our criterion for nondominance is that we've
never established that an individual is dominated on an
objective, and because that's a positive comparison, it
never evaluates True for NaN.  This is entirely consistent
with the expected behavior of NaN.  NaN is never greater
than, less than, or equal to any other number.

The end result is that the more NaNs an individual has in
its objectives, the more likely it is to be nondominated.
This culminates in the ultimate failure, which is that an
individual with all NaN objectives dominates any other
individual.  Adding a NaN check means that real-world
broken models that produce NaNs in some parts of the
decision space will produce individuals that sort worse
than individuals in the non-broken part of the space.

We might want to find a way to issue a warning.