Sampling (U): Accelerate issued check with a set
This "cheats" a little bit by adding a size-mutable data
structure.  It goes against my rule that I want this to be
translatable to C with a minimum of fuss.  However, it's
accelerating a linear scan that would be way faster in C
and therefore wouldn't benefit nearly as much from hashing.
I.e. we could do without it just fine in C.  Furthermore,
it gives us two orders of magnitude speed improvement in
Python, when we have an issue list we allow to grow to
10k in size.