COPYING: Change license to 3-clause BSD from GPL3
The reason we're open-sourcing this project is that we want
to have lots of people using MOEAs.  The known alternatives
are very academic in orientation and are generally designed
around running test problems.  Furthermore their licensing
is either written ad-hoc by the academics who developed
them (a red flag to lawyers) or it is copyleft (GPL).  This
combination of licensing uncertainty and the need for source
modification thrusts many users into a legal grey area
and a technical morass.

This is why we're switching to a permissive license: the
3-clause BSD license allows anybody to use and modify
the source code, and to incorporate it in any project,
even proprietary ones.  We view this as the strongest
statement we can make that we want people using MOEAs.
We believe in them and we think people who use them will
make better decisions.

And we think people might be willing to pay us for:

(1) Consulting services
(2) Translating the project into different languages.
(3) Integrating δMOEA into other programs.